requirements.txt
accelerate==1.6.0
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.51
langchain_experimental==0.3.4
langchain_milvus==0.1.9
langchain_ollama==0.3.1
pymilvus==2.5.5
pymilvus_model==0.3.2
pypdf==5.4.0
python-dotenv==1.1.0
sentence_transformers==4.0.2
tqdm==4.66.5
